add L14

+++ L14/0.py
@@ -1,8 +1,15 @@
-from pycat.core import Color, Sprite, Window, Scheduler
+from pycat.core import Color, Sprite, Window, Scheduler, Label
 from random import randint
 
 window = Window()
 
+
+class Score(Label):
+    def on_create(self):
+        self.text = "0"
+        self.score = 0
+    def on_update(self, dt):
+        self.text = str(self.score)
 
 class Player(Sprite):
 
@@ -50,6 +57,7 @@ class PowerPlayerBullet(Sprite):
         self.position = player.position
         self.speed = 10
         self.point_toward_mouse_cursor()
+        self.add_tag("pbullet")
     def on_update(self, dt):
         self.move_forward(self.speed)
         if self.is_touching_window_edge():
@@ -62,15 +70,46 @@ class Enemy(Sprite):
         self.speed = 3
         self.scale = 30
         self.color = Color.CHARTREUSE
+        self.t = 0
+        self.hp = 20
+        self.enemy_time = 0
     def on_update(self,dt):
         self.move_forward(self.speed)
+        self.enemy_time += dt
+        if self.is_touching_any_sprite_with_tag("pbullet"):
+            if self.hp > 0:
+                self.hp -= 5
+                self.opacity -= 30
+            else:
+                self.delete()
         if self.is_touching_window_edge():
-            self.delete()
+            if self.t <= 2:
+                self.t += dt
+            else:
+                self.delete()
+        if self.enemy_time > 2:
+            bullet = window.create_sprite(EnemyBullet)
+            bullet.position = self.position
+            bullet.point_toward_sprite(player)
+            self.enemy_time = 0
+
+
 
 def spawn_enemy():
     window.create_sprite(Enemy)
 
+class EnemyBullet(Sprite):
+    def on_create(self):
+        self.scale = 10
+        self.color = Color.CYAN
+    def on_update(self, dt):
+        self.move_forward(5)
+
+#haven't added damage to player yet
+
+
 Scheduler.update(spawn_enemy, delay=2)
 
+score = window.create_label(Score)
 player = window.create_sprite(Player)
 window.run()
